+++ web/src/buildInfo.ts
@@ -3,9 +3,15 @@ declare const __BUILD_NUMBER__: string;
 declare const __GIT_SHA__: string;
 declare const __BUILD_TIME__: string;
 declare const __RELEASE_URL__: string;
+declare const __APP_BRAND_NAME__: string;
+declare const __APP_BRAND_URL__: string;
+declare const __APP_BRAND_LOGO__: string;
 
 export const appVersion = __APP_VERSION__;
 export const buildNumber = __BUILD_NUMBER__;
 export const gitSha = __GIT_SHA__;
 export const buildTime = __BUILD_TIME__;
 export const releaseURL = __RELEASE_URL__;
+export const appBrandName = __APP_BRAND_NAME__;
+export const appBrandURL = __APP_BRAND_URL__;
+export const appBrandLogo = __APP_BRAND_LOGO__;